Namespace fixes.
Includes the following fixes for PSR namespace autoload support:

* Add missing use shortcuts.
* Separate out AliceBuilder from AliceAxolotlParameters.
* Use __construct instead of class named constructor in
    UnacknowledgedPreKeyMessageItems.
