+++ src/state/SessionState.php
@@ -407,7 +407,7 @@ class SessionState
 }
 class UnacknowledgedPreKeyMessageItems
 {
-    public function UnacknowledgedPreKeyMessageItems($preKeyId, $signedPreKeyId, $baseKey)
+    public function __construct($preKeyId, $signedPreKeyId, $baseKey)
     {
         /*
         :type preKeyId: int
